Default

+++ onmt/Loss.py
@@ -246,7 +246,7 @@ class NMTLossCompute(LossComputeBase):
         loss_data_unk = loss_unk.data.clone()
         loss_data = loss.data.clone()
         LOSS =loss+loss_unk
-        stats = self._stats(loss_data_unk, loss_data, scores_unk.data, scores.data, target_unk.view(-1).data*target_unk.data.ne(self.unk_idx), target.view(-1).data*target_unk.data.eq(self.unk_idx))
+        stats = self._stats(loss_data_unk, loss_data, scores_unk.data, scores.data, target_unk.view(-1).data*target_unk.data.ne(self.unk_idx).int(), target.view(-1).data*target_unk.data.eq(self.unk_idx).int())
 
         return LOSS, stats
